Move DataMigrationError enum to separate file
